+++ pysd/builders/python/namespace.py
