+++ pysd/cli/main.py
@@ -8,9 +8,9 @@ from datetime import datetime
 from .parser import parser
 
 import pysd
-from pysd.translation.vensim.vensim_utils import supported_extensions as\
+from pysd.translators.vensim.vensim_utils import supported_extensions as\
     vensim_extensions
-from pysd.translation.xmile.xmile_utils import supported_extensions as\
+from pysd.translators.xmile.xmile_utils import supported_extensions as\
     xmile_extensions
