+++ pysd/pysd.py
@@ -61,8 +61,8 @@ def read_xmile(xmile_file, data_files=None, initialize=True,
     >>> model = read_xmile('../tests/test-models/samples/teacup/teacup.xmile')
 
     """
-    from pysd.translation.xmile.xmile_file import XmileFile
-    from pysd.building.python.python_model_builder import ModelBuilder
+    from pysd.translators.xmile.xmile_file import XmileFile
+    from pysd.builders.python.python_model_builder import ModelBuilder
 
     # Read and parse Xmile file
     xmile_file_obj = XmileFile(xmile_file)
@@ -138,8 +138,8 @@ def read_vensim(mdl_file, data_files=None, initialize=True,
     >>> model = read_vensim('../tests/test-models/samples/teacup/teacup.mdl')
 
     """
-    from pysd.translation.vensim.vensim_file import VensimFile
-    from pysd.building.python.python_model_builder import ModelBuilder
+    from pysd.translators.vensim.vensim_file import VensimFile
+    from pysd.builders.python.python_model_builder import ModelBuilder
     # Read and parse Vensim file
     ven_file = VensimFile(mdl_file, encoding=encoding)
     ven_file.parse()
