+++ pysd/translators/vensim/parsing_grammars/element_object.peg
