+++ pysd/translators/vensim/vensim_section.py
